Core/Quests: Implemented QUEST_OBJECTIVE_FLAG_2_QUEST_BOUND_ITEM that makes required items not be stored in inventory
Port From (https://github.com/TrinityCore/TrinityCore/commit/4dce6ab228c54319032b346d2d791262eac3d9cd)

+++ Source/Game/Networking/Packets/QuestPackets.cs
@@ -227,7 +227,7 @@ namespace Game.Networking.Packets
                     _worldPacket.WriteInt32(questObjective.ObjectID);
                     _worldPacket.WriteInt32(questObjective.Amount);
                     _worldPacket.WriteUInt32((uint)questObjective.Flags);
-                    _worldPacket.WriteUInt32(questObjective.Flags2);
+                    _worldPacket.WriteUInt32((uint)questObjective.Flags2);
                     _worldPacket.WriteFloat(questObjective.ProgressBarWeight);
 
                     _worldPacket.WriteInt32(questObjective.VisualEffects.Length);
